remove newline
nevermind, this is way stupider. Apparently the links break if you have
some spacing after the begining of the list.

update:
After som experimenting, my conclusion is that 4 spaces is what triggers
the verbatim parsing, but it only happens if there is a newline before
the textblock? And apparently html-tags is ignored in this check
(probably due to the library used to parse markdown)

wth