Jumfile and compose, and more cleanup (#540)
* add /en to jumpfile

and change tabs to spaces

* change port env name in compose file

* remove en titles from meta.toml for pages without en body

they still show up in the sidebar now, but they lead to an error page

* change indent type on ior page

this was breaking the links, since indentation with enough spaces in
markdown (8 in this case maybe? I think it varies by which flavour of
markdown) means that the text should be parsed verbatim.

* remove newline

nevermind, this is way stupider. Apparently the links break if you have
some spacing after the begining of the list.

update:
After som experimenting, my conclusion is that 4 spaces is what triggers
the verbatim parsing, but it only happens if there is a newline before
the textblock? And apparently html-tags is ignored in this check
(probably due to the library used to parse markdown)

wth

* remove outdated current projects section

it was already removed on the swedish page